+++ OpenPlottingLibrary/OpenPlottingLibrary/Point3D.cs
@@ -4,9 +4,9 @@
     {
         public float x { get; private set; }
         public float y { get; private set; }
-        public float z { get; private set; }
+        public double z { get; private set; }
 
-        public Point3D(float x, float y, float z) : this()
+        public Point3D(float x, float y, double z) : this()
         {
             this.x = x;
             this.y = y;
